Image Detection Task/query_detr_minimal/requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
pillow>=9.5
matplotlib>=3.7
tqdm>=4.65
pycocotools>=2.0.7
scipy>=1.10
